project/requirements.txt
albumentations==0.4.6
appnope==0.1.0
attrs==19.3.0
backcall==0.2.0
bleach==3.1.5
certifi==2020.6.20
chardet==3.0.4
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
entrypoints==0.3
future==0.18.2
idna==2.10
imageio==2.9.0
imgaug==0.4.0
importlib-metadata==1.7.0
jedi==0.17.2
Jinja2==2.11.2
json5==0.9.5
jsonschema==3.2.0
jupyterlab==2.2.0
kaggle==1.5.6
MarkupSafe==1.1.1
mistune==0.8.4
nbconvert==5.6.1
nbformat==5.0.7
networkx==2.4
notebook==6.0.3
opencv-python==4.3.0.36
opencv-python-headless==4.3.0.36
packaging==20.4
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
progress==1.5
ptyprocess==0.6.0
Pygments==2.6.1
pyparsing==2.4.7
pyrsistent==0.16.0
python-dateutil==2.8.1
python-slugify==4.0.1
pytz==2020.1
PyWavelets==1.1.1
PyYAML==5.3.1
pyzmq==19.0.1
requests==2.24.0
Send2Trash==1.5.0
Shapely==1.7.0
six==1.15.0
terminado==0.8.3
testpath==0.4.4
text-unidecode==1.3
torch==1.4.0
torchvision==0.5.0
tornado==6.0.4
tqdm==4.48.0
traitlets==4.3.3
urllib3==1.24.3
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
